refactor: version, imports, types
